+++ kadai3-1/decoch/main.go
@@ -1,91 +1,31 @@
 package main
 
 import (
-	"bufio"
 	"fmt"
-	"io/ioutil"
-	"log"
-	"math/rand"
-	"os"
-	"strings"
-	"sync"
-	"time"
+
+	"github.com/decoch/dojo4/kadai3-1/decoch/score"
+	"github.com/decoch/dojo4/kadai3-1/decoch/timer"
+	"github.com/decoch/dojo4/kadai3-1/decoch/typing"
 )
 
 func main() {
-	fmt.Println("### typing game start. ###")
+	fmt.Println("### start typing game. ###")
 
-	score := 0
+	typeCh := typing.Words()
+	timerCh := timer.NewCh(10)
 
-	ch1 := input()
-	ch2 := timer()
+	counter := new(score.Counter)
 
-	var m sync.Mutex
-	select {
-	case v1 := <-ch1:
-		m.Lock()
-		if v1 {
-			score++
-		}
-		m.Unlock()
-	case _ = <-ch2:
-		m.Lock()
-		fmt.Printf("\n### Score: %d ###\n", score)
-		fmt.Println("### typing game end. ###")
-		m.Unlock()
-	}
-}
-
-func input() <-chan bool {
-	inputCh := make(chan bool)
-	go func() {
-		words, err := getWords()
-		if err != nil {
-			log.Fatalln("cannot get words.", err)
-		}
-
-		stdin := bufio.NewScanner(os.Stdin)
-		rand.Seed(time.Now().UnixNano())
-		for {
-			position := rand.Intn(len(words))
-			word := words[position]
-			fmt.Println(word)
-
-			stdin.Scan()
-			input := stdin.Text()
-			if word == input {
-				fmt.Printf("### GOT IT!!!! ###\n\n")
-				inputCh <- true
-			} else {
-				fmt.Printf("### WRONG ###\n\n")
-				inputCh <- false
+	for {
+		select {
+		case v1 := <-typeCh:
+			if v1 {
+				counter.Add(1)
 			}
+		case _ = <-timerCh:
+			fmt.Printf("\n### Score: %d ###\n", counter.Value())
+			fmt.Println("### end typing game. ###")
+			return
 		}
-		close(inputCh)
-	}()
-	return inputCh
-}
-
-func timer() <-chan struct{} {
-	timerCh := make(chan struct{})
-	go func() {
-		time.Sleep(2 * time.Second)
-		timerCh <- struct{}{}
-		close(timerCh)
-	}()
-	return timerCh
-}
-
-func getWords() ([]string, error) {
-	data, err := readFile()
-	if err != nil {
-		return nil, err
 	}
-	samples := strings.Split(strings.TrimSpace(string(data)), "\n")
-	return samples, nil
-}
-
-func readFile() ([]byte, error) {
-	data, err := ioutil.ReadFile("./data/word.txt")
-	return data, err
 }
