+++ kadai3-1/decoch/score/counter.go
@@ -1,0 +1,20 @@
+package score
+
+import "sync"
+
+type Counter struct {
+	value int
+	m     sync.Mutex
+}
+
+func (s *Counter) Add(v int) {
+	s.m.Lock()
+	defer s.m.Unlock()
+	s.value = s.value + v
+}
+
+func (s *Counter) Value() int {
+	s.m.Lock()
+	defer s.m.Unlock()
+	return s.value
+}
